+++ pycantonese/data/rime_cantonese/download_and_resegment.py
@@ -21,7 +21,7 @@ except ModuleNotFoundError:
     )
 
 
-_BRANCH = "2020.09.09"
+_BRANCH = "2021.05.16"
 _URL = f"https://github.com/jacksonllee/rime-cantonese/archive/{_BRANCH}.zip"
 
 _THIS_DIR = os.path.abspath(os.path.dirname(__file__))