+++ pycantonese/jyutping/characters.py
@@ -1,10 +1,11 @@
 from collections import Counter, defaultdict
 from functools import lru_cache
+from typing import List, Tuple
 
 from pycantonese.corpus import hkcancor, Token
 from pycantonese.data.rime_cantonese import CHARS_TO_JYUTPING, LETTERED
 from pycantonese.jyutping.parse_jyutping import parse_jyutping
-from pycantonese.word_segmentation import segment
+from pycantonese.word_segmentation import _get_default_segmenter, segment, Segmenter
 from pycantonese.util import _split_chars_with_alphanum, _deprecate
 
 
@@ -35,70 +36,60 @@ def _get_words_characters_to_jyutping():
     for word, jyutping_counter in words_to_jyutping_counters.items():
         jp = jyutping_counter.most_common(1)[0][0]
         words_to_jyutping[word] = jp
-    characters_to_jyutping = {}
+    chars_to_jp = {}
     for character, jyutping_counter in characters_to_jyutping_counters.items():
         jp = jyutping_counter.most_common(1)[0][0]
-        characters_to_jyutping[character] = jp
+        chars_to_jp[character] = jp
 
     words_to_jyutping = {
-        # The ordering of the following dicts matters. The rime-cantonese
-        # data may contain what's been re-segmented by this repo, and may
-        # contain jyutping pronunciations for particular characters that
-        # are only used in those contexts. The data from HKCanCor should comes
-        # last to act as the default to override such cases.
+        # The ordering of the following dicts matters.
+        # rime-cantonese (more accurate data) overrides HKCanCor if they don't agree.
+        **words_to_jyutping,
         **{k: v for k, v in LETTERED.items() if len(_split_chars_with_alphanum(k)) > 1},
         **{k: v for k, v in CHARS_TO_JYUTPING.items() if len(k) > 1},
-        **words_to_jyutping,
     }
 
     # TODO: Extract characters from CHARS_TO_JYUTPING and LETTERED
     #    and add them to characters_to_jyutping
-    characters_to_jyutping = {
-        # The ordering of the following dicts matters. The rime-cantonese
-        # data may contain what's been re-segmented by this repo, and may
-        # contain jyutping pronunciations for particular characters that
-        # are only used in those contexts. The data from HKCanCor should comes
-        # last to act as the default to override such cases.
+    chars_to_jp = {
+        # The ordering of the following dicts matters.
+        # rime-cantonese (more accurate data) overrides HKCanCor if they don't agree.
+        **chars_to_jp,
         **{k: v for k, v in LETTERED.items() if len(k) == 1},
         **{k: v for k, v in CHARS_TO_JYUTPING.items() if len(k) == 1},
-        **characters_to_jyutping,
     }
 
-    return words_to_jyutping, characters_to_jyutping
+    return words_to_jyutping, chars_to_jp
 
 
-def characters_to_jyutping(chars, segmenter=None):
+def characters_to_jyutping(
+    chars: str, segmenter: Segmenter = None
+) -> List[Tuple[str, str]]:
     """Convert Cantonese characters into Jyutping romanization.
 
     The conversion model is based on the HKCanCor corpus and rime-cantonese
     data. Any unseen Cantonese character (or punctuation mark,
     for that matter) is represented by `None` in the output.
 
-    The output is a list of segmented words, where each word is a 2-tuple of
-    (Cantonese characters, Jyutping romanization).
-
-    .. versionadded:: 3.0.0
-        This function replaces the deprecated equivalent
-        ``characters2jyutping``.
-
-    .. versionchanged:: 3.0.0
-        The returned valued is now a list of segmented words,
-        where each word is a 2-tuple of (Cantonese characters, Jyutping).
-        Previously, it was a list of Jyutping strings for the individual
-        Cantonese characters.
+    This function also performs word segmentation, in order to resolve potential
+    ambiguity in mapping characters to Jyutping.
 
     Parameters
     ----------
     chars : str
         A string of Cantonese characters.
-    segmenter : :class:`~pycantonese.word_segmentation.Segmenter`, optional
-        If ``None`` (default value) or not provided, the default word segmenter
-        is used. For custom behavior, pass in a segmenter with specified allowed and/or
-        disallowed potential words.
+    segmenter : Segmenter, optional
+        A :class:`~pycantonese.word_segmentation.Segmenter` instance to customize
+        word segmentation.
+        If specified, this segmenter is passed to the ``cls`` keyword argument of
+        :func:`~pycantonese.segment`.
+        If ``None`` or not given, the default segmenter is used.
 
     Returns
     -------
-    list[tuple[str]]
+    list[tuple[str, str]]
+        A list of segmented words, where each word is a 2-tuple of
+        (Cantonese characters, Jyutping romanization).
 
     Examples
     --------
@@ -109,6 +100,8 @@ def characters_to_jyutping(chars, segmenter=None):
         return []
     words_to_jyutping, chars_to_jyutping = _get_words_characters_to_jyutping()
     result = []
+    if segmenter is None:
+        segmenter = _get_default_segmenter()
     for word in segment(chars, cls=segmenter):
         try:
             jp = words_to_jyutping[word]